feat: Add HomeWidget and SummaryCard, refactor HomeScreen to use new components and remove HomeNavbarWidget
