+++ lib/widgets/catat_cepat.dart
@@ -1,0 +1,94 @@
+import 'package:flutter/material.dart';
+import 'package:finlog/styles/colors.dart';
+
+class CatatCepat extends StatefulWidget {
+  const CatatCepat({super.key});
+
+  @override
+  State<CatatCepat> createState() => _CatatCepatState();
+}
+
+class _CatatCepatState extends State<CatatCepat> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(20.0),
+      decoration: BoxDecoration(
+        gradient: const LinearGradient(
+          colors: [finlogBluePrimaryDark, finlogBluePrimary],
+          begin: Alignment.topLeft,
+          end: Alignment.bottomRight,
+        ),
+        borderRadius: BorderRadius.circular(16),
+        boxShadow: [
+          BoxShadow(
+            color: finlogBluePrimary.withOpacity(0.3),
+            blurRadius: 8,
+            offset: const Offset(0, 4),
+          ),
+        ],
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Row(
+            children: const [
+              Text(
+                'Catat Cepat',
+                style: TextStyle(
+                  fontSize: 18,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.white,
+                ),
+              ),
+              SizedBox(width: 8),
+              Icon(Icons.flash_on, color: Colors.white, size: 20),
+            ],
+          ),
+          const SizedBox(height: 12),
+          Text(
+            'Nominal',
+            style: TextStyle(
+              color: Colors.white.withOpacity(0.8),
+              fontSize: 13,
+            ),
+          ),
+          const SizedBox(height: 6),
+          TextField(
+            controller: TextEditingController(text: '1,000,000'), // Placeholder
+            style: const TextStyle(
+              color: Colors.black87,
+              fontSize: 16,
+              fontWeight: FontWeight.w500,
+            ),
+            keyboardType: TextInputType.number,
+            decoration: InputDecoration(
+              filled: true,
+              fillColor: Colors.white,
+              contentPadding: const EdgeInsets.symmetric(
+                horizontal: 16,
+                vertical: 12,
+              ),
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(10),
+                borderSide: BorderSide.none,
+              ),
+            ),
+          ),
+          const SizedBox(height: 12),
+          Text(
+            'Catat dengan\ncepat dan mudah~\nketuk disini!',
+            style: TextStyle(
+              color: Colors.white.withOpacity(0.9),
+              fontSize: 14,
+              height: 1.4,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+
+// TODO : implement widget to accept the user money input
